requirements.txt
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-dysmsapi==1.0.0
Django==2.0.6
django-shortuuidfield==0.1.3
mysqlclient==1.3.12
Pillow==5.1.0
PyMySQL==0.8.1
pytz==2018.4
redis==2.10.6
shortuuid==0.5.0
six==1.11.0
